pusher/requirements.txt
aio-pika==9.5.8
aiormq==6.9.2
aiosmtplib==5.0.0
exceptiongroup==1.3.1
idna==3.11
multidict==6.7.0
pamqp==3.3.0
propcache==0.4.1
python-dotenv==1.2.1
typing_extensions==4.15.0
yarl==1.22.0
